Create calculate-distance file to store the distance calculating function and import the function in the truck card
